Split settings.rs into settings/ subdirectory
Break the 2217-line settings.rs into five focused files under
settings/:

  settings/mod.rs      — submodule decls + pub(crate) re-exports
  settings/cli.rs      — Cli struct + typed enums (MissingFontsArg,
                         LogLevel, DataAccessMode, LogFormatArg) + 4
                         CLI-parsing tests
  settings/env.rs      — ENV_* constants + SETTING_PAIRS +
                         EnvValues::from_env (includes the PaaS PORT
                         fallback)
  settings/parsers.rs  — InputKind + field_name + every parse_*
                         helper, @file/stdin loading, JSON/locale/
                         google-font parsers + 6 parser tests
  settings/resolve.rs  — ResolvedSettings + Overrides +
                         BootstrapOverrides + WorkingSettings +
                         resolve_settings + parse_{env,cli}_overrides +
                         resolve_bootstrap + load_converter_config +
                         data-access helpers + 13 end-to-end tests
                         (with their own EnvGuard + ENV_LOCK)

Public surface unchanged: main.rs still does `use
settings::{resolve_settings, Cli}` and `resolve.rs` keeps
ResolvedSettings as pub(crate) (accessed by main.rs through type
inference). Typed CLI enums had to be promoted from pub(super) to
pub(crate) to match Cli's own visibility.

All 23 settings tests + the rest of the crate (33 unit, 139 integration)
pass.
